+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 IMPROVEMENTS:
 
+  * **New resource: `aws_sqs_queue`** [GH-1939]
   * provider/aws: support ec2 termination protection [GH-1988]
 
 ## 0.5.2 (May 15, 2015)